requirements.txt
torch
numpy
pandas
tifffile
Pillow
scikit-image
scipy
tqdm
matplotlib
